@@ -34,27 +34,27 @@
     "release:changelog": "npm run conventional-changelog -- -p angular -i CHANGELOG.md -s"
   },
   "dependencies": {
-    "@agm/core": "1.0.0-beta.2",
-    "@angular/animations": "6.0.0",
-    "@angular/cdk": "^6.4.7",
-    "@angular/common": "6.0.0",
-    "@angular/compiler": "6.0.0",
-    "@angular/core": "6.0.0",
-    "@angular/forms": "6.0.0",
-    "@angular/http": "6.0.0",
-    "@angular/platform-browser": "6.0.0",
-    "@angular/platform-browser-dynamic": "6.0.0",
-    "@angular/router": "6.0.0",
+    "@agm/core": "^1.0.0-beta.5",
+    "@angular/animations": "^7.0.3",
+    "@angular/cdk": "^7.0.3",
+    "@angular/common": "^7.0.3",
+    "@angular/compiler": "^7.0.3",
+    "@angular/core": "^7.0.3",
+    "@angular/forms": "^7.0.3",
+    "@angular/http": "^7.0.3",
+    "@angular/platform-browser": "^7.0.3",
+    "@angular/platform-browser-dynamic": "^7.0.3",
+    "@angular/router": "^7.0.3",
     "@asymmetrik/ngx-leaflet": "3.0.1",
     "@nebular/auth": "2.0.1",
     "@nebular/bootstrap": "2.0.1",
     "@nebular/security": "2.0.1",
     "@nebular/theme": "2.0.1",
-    "@ng-bootstrap/ng-bootstrap": "1.0.0",
-    "@swimlane/ngx-charts": "7.0.1",
+    "@ng-bootstrap/ng-bootstrap": "^4.0.0",
+    "@swimlane/ngx-charts": "^10.0.0",
     "angular-tree-component": "7.2.0",
     "angular2-chartjs": "0.4.1",
-    "angular2-toaster": "6.0.0",
+    "angular2-toaster": "^6.1.0",
     "bootstrap": "4.0.0",
     "chart.js": "2.7.1",
     "ckeditor": "4.7.3",
@@ -65,15 +65,15 @@
     "ionicons": "2.0.1",
     "leaflet": "1.2.0",
     "nebular-icons": "1.0.9",
-    "ng2-ckeditor": "1.1.13",
+    "ng2-ckeditor": "^1.2.2",
     "ng2-completer": "2.0.8",
     "ng2-smart-table": "1.3.5",
-    "ngx-echarts": "2.0.1",
+    "ngx-echarts": "^4.0.1",
     "normalize.css": "6.0.0",
     "pace-js": "1.0.2",
     "roboto-fontface": "0.8.0",
-    "rxjs": "^6.1.0",
-    "rxjs-compat": "^6.1.0",
+    "rxjs": "6.3.0",
+    "rxjs-compat": "6.3.0",
     "socicon": "3.0.5",
     "tinymce": "4.5.7",
     "typeface-exo": "0.0.22",
@@ -81,18 +81,19 @@
     "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "6.0.0",
-    "@angular/compiler-cli": "6.0.0",
-    "@angular/language-service": "6.0.0",
+    "@angular-devkit/build-angular": "0.10.2",
+    "@angular/compiler-cli": "^7.0.3",
+    "@angular/cli": "^7.0.3",
+    "@angular/language-service": "7.0.0",
     "@compodoc/compodoc": "1.0.1",
     "@fortawesome/fontawesome-free": "^5.2.0",
     "@types/d3-color": "1.0.5",
-    "@types/googlemaps": "3.30.4",
+    "@types/googlemaps": "^3.30.4",
     "@types/jasmine": "2.5.54",
     "@types/jasminewd2": "2.0.3",
     "@types/leaflet": "1.2.3",
     "@types/node": "6.0.90",
-    "codelyzer": "4.0.2",
+    "codelyzer": "^4.5.0",
     "conventional-changelog-cli": "1.3.4",
     "husky": "0.13.3",
     "jasmine-core": "2.6.4",
@@ -109,8 +110,7 @@
     "stylelint": "7.13.0",
     "ts-node": "3.2.2",
     "tslint": "5.7.0",
-    "tslint-language-service": "0.9.6",
-    "typescript": "2.7.2",
-    "@angular-devkit/build-angular": "~0.6.0"
+    "tslint-language-service": "^0.9.9",
+    "typescript": "3.1.3"
   }
 }
