@@ -2,8 +2,8 @@ import { Component, OnDestroy } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { combineLatest } from 'rxjs';
 import { takeWhile } from 'rxjs/operators';
-import { NgxEchartsService } from 'ngx-echarts';
 import { NbThemeService } from '@nebular/theme';
+import { registerMap } from 'echarts';
 
 @Component({
   selector: 'ngx-bubble-map',
@@ -31,8 +31,7 @@ export class BubbleMapComponent implements OnDestroy {
   private alive = true;
 
   constructor(private theme: NbThemeService,
-              private http: HttpClient,
-              private es: NgxEchartsService) {
+              private http: HttpClient) {
 
     combineLatest([
       this.http.get('assets/map/world.json'),
@@ -41,7 +40,7 @@ export class BubbleMapComponent implements OnDestroy {
       .pipe(takeWhile(() => this.alive))
       .subscribe(([map, config]: [any, any]) => {
 
-        this.es.registerMap('world', map);
+        registerMap('world', map);
 
         const colors = config.variables;
         this.bubbleTheme = config.variables.bubbleMap;
